requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-login
flask-wtf
email-validator
requests
pyOpenSSL
gunicorn